requirements.txt
amqp==5.0.6
asgiref==3.3.4
attrs==21.2.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dj-rest-auth==2.1.5
Django==3.2.3
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-extensions==3.1.3
django-jazzmin==2.4.7
django-timezone-field==4.1.2
djangorestframework==3.12.4
drf-spectacular==0.16.0
flake8==3.9.2
idna==2.10
inflection==0.5.1
isort==5.8.0
jsonschema==3.2.0
kombu==5.0.2
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
prompt-toolkit==3.0.18
pycodestyle==2.7.0
pyflakes==2.3.1
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
yapf==0.31.0
